[No-Jira] Accept missing instance progress poppup as valid state of the application
There are some non deterministic situations in our application, when some popover dialog may or may not be present on the page.
Examples of such situations are:
1.) Consent with cookies popover in the very start of the application/test
2.) Smart Event Instance status progress popover

The 1.) is for some user displayed, for some it is not. We Are not sure why this happens. The guess for a root case is a country of the machine, where the test is running. Needs more investigations/test results tracking.
The 2.) displays progress as three steps. However the time combination of time needed for the operation on backend with polling interval from UI causes, soemtimes the progress update is not 'smooth' and popover disappears faster than expected in the test.

*This PR tries to stabilize 2.)*

The error handling was addopted according to th ecyprtess example:
- https://github.com/cypress-io/cypress-example-recipes/blob/master/examples/fundamentals__errors/cypress/e2e/test-fails.cy.js